requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
distro==1.9.0
exceptiongroup==1.2.1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
immutabledict==4.2.0
jsonpatch==1.33
jsonpointer==2.4
langchain==0.2.3
langchain-core==0.2.5
langchain-text-splitters==0.2.1
langsmith==0.1.75
multidict==6.0.5
numpy==1.26.4
openai==1.33.0
orjson==3.10.3
ortools==9.10.4067
packaging==23.2
pandas==2.2.2
praw==7.7.1
prawcore==2.4.0
protobuf==5.27.1
pydantic==2.7.3
pydantic_core==2.18.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
tenacity==8.3.0
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
update-checker==0.18.0
urllib3==2.2.1
websocket-client==1.8.0
yarl==1.9.4
langchain_openai==1.0.0
